refactor: move utils and correct test setup
